% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPhenoData2ExposomeSetDS.R
\name{addPhenoData2ExposomeSetDS}
\alias{addPhenoData2ExposomeSetDS}
\title{Add Phenotype data to ExposomeSet}
\usage{
addPhenoData2ExposomeSetDS(
  x,
  pheno,
  identifier_ExposomeSet,
  identifier_new_phenotypes,
  complete_cases
)
}
\arguments{
\item{x}{\code{ExposomeSet} ExposomeSet to which add phenotype information}

\item{pheno}{\code{data.frame} Table with the new phenotypes}

\item{identifier_ExposomeSet}{\code{character} Name of the ID column on the already present 
phenotypes data.frame on the ExposomeSet}

\item{identifier_new_phenotypes}{\code{character} Name of the ID column on the new phenotypes data.frame}

\item{complete_cases}{\code{bool} If \code{TRUE} only the matching individuals 
between the ExpressionSet and the phenotypes table will be included on the resulting ExpressionSet. If 
\code{FALSE} all the individuals on the input ExpressionSet will be on the output ExpressionSet}
}
\value{
\code{ExposomeSet} with updated phenotypes information
}
\description{
Add phenotype data to an ExpressionSet
}
