% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anderson.darling.testDS.R
\name{anderson.darling.testDS}
\alias{anderson.darling.testDS}
\title{Anderson-Darling Normality Test}
\usage{
anderson.darling.testDS(x)
}
\arguments{
\item{x}{\code{numeric} Name of the numeric vector}
}
\value{
Results of the Anderson-Darling Normality Test
}
\description{
Perform the Anderson-Darling Normality Test on a numeric vector (or column of a data frame)
}
