% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effective.testsDS.R
\name{effective.testsDS}
\alias{effective.testsDS}
\title{Obtain the threshold for effective tests of an ExposomeSet}
\usage{
effective.testsDS(set)
}
\arguments{
\item{set}{\code{ExposomeSet} Exposome Set object}
}
\value{
\code{numeric} threshold for effective tests
}
\description{
Obtain the threshold for effective tests of an ExposomeSet
}
