% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposome_pca_pooledDS.R
\name{exposome_pca_pooled_addPCDS}
\alias{exposome_pca_pooled_addPCDS}
\title{Title}
\usage{
exposome_pca_pooled_addPCDS(object, pca)
}
\arguments{
\item{object}{}

\item{pc}{}
}
\value{

}
\description{
Title
}
