% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposome_pca_pooledDS.R
\name{exposome_pca_pooledDS}
\alias{exposome_pca_pooledDS}
\title{Get block SVD}
\usage{
exposome_pca_pooledDS(object)
}
\arguments{
\item{Set}{\code{ExposomeSet}}
}
\value{
\code{data.frame} with block results
}
\description{
Calculate block SVD decomposition
}
