% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposome_pcaDS.R
\name{exposome_pcaDS}
\alias{exposome_pcaDS}
\title{Principal components analysis of an Exposome Set}
\usage{
exposome_pcaDS(object)
}
\arguments{
\item{object}{\code{ExposomeSet} Exposome Set object}
}
\value{
Returns a \code{ExposomeSet PCA} object
}
\description{
Performs a non-disclosive PCA given an Exposome Set, the Exposome Set can be subsetted by families to 
perform the PCA
}
