% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposome_scale_exposuresDS.R
\name{exposome_scale_exposuresDS}
\alias{exposome_scale_exposuresDS}
\title{Title}
\usage{
exposome_scale_exposuresDS(Set, means)
}
\arguments{
\item{Set}{}

\item{means}{}
}
\value{

}
\description{
Title
}
