% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposomeSubsetDS.R
\name{exposomeSubsetDS}
\alias{exposomeSubsetDS}
\title{Subset Exposome Set by familie(s)}
\usage{
exposomeSubsetDS(set, fam)
}
\arguments{
\item{set}{\code{ExposomeSet} Exposome Set object}

\item{fam}{\code{character vector} Families to subset the exposome dataset}
}
\value{
Subsetted \code{ExposomeSet}
}
\description{
Subset Exposome Set by familie(s)
}
