% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposureNamesDS.R
\name{exposureNamesDS}
\alias{exposureNamesDS}
\title{Exposure Names}
\usage{
exposureNamesDS(exposomeSet)
}
\arguments{
\item{exposomeSet}{\code{ExposomeSet} Exposome Set object}
}
\value{
\code{character vector} of the exposure names
}
\description{
Extract the exposure names from a Expression Set object
}
