% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposures_pData.R
\name{exposures_pData}
\alias{exposures_pData}
\title{Exposures and phenotyphes data}
\usage{
exposures_pData(exposomeSet)
}
\arguments{
\item{exposomeSet}{\code{ExposomeSet} Exposome Set object}
}
\value{
\code{data.frame} With exposures and phenotypes
}
\description{
Extract a table with the exposures and phenotypes data for all the individuals of an
Exposome Set
}
