% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractExposomeSetFromImputedSetDS.R
\name{extractExposomeSetFromImputedSetDS}
\alias{extractExposomeSetFromImputedSetDS}
\title{Extract ExposomeSet from imExposomeSet}
\usage{
extractExposomeSetFromImputedSetDS(x, rid)
}
\arguments{
\item{x}{\code{imExposomeSet} to which extract one imputed ExposomeSet.}

\item{rid}{\code{numeric} Number of imputed set to extract.}
}
\value{
\code{ExposomeSet}
}
\description{
Extract one of the imputed sets of an imputed ExposomeSet (imExposomeSet)
}
