% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familyNamesDS.R
\name{familyNamesDS}
\alias{familyNamesDS}
\title{Get family names from Exposome Set}
\usage{
familyNamesDS(object, by.exposure = FALSE)
}
\arguments{
\item{object}{\code{ExposomeSet} Exposome Set object}

\item{by.exposure}{\code{bool} (default \code{FALSE}) If \code{TRUE} a vector labeled with each exposure 
name will be returned with the family of each exposures. If \code{FALSE} a vector with the 
(unique) name of the families of exposures will be returned.}
}
\value{
A \code{character vector} with the family names of the Exposome Set
}
\description{
Get family names from Exposome Set
}
