% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_clinical_exposure_by_geoDS.R
\name{get_exposure_from_geoDS}
\alias{get_exposure_from_geoDS}
\title{Title}
\usage{
get_exposure_from_geoDS(
  lat,
  lon,
  exposure,
  exposure_name,
  clinical,
  clinical_lat_var,
  clinical_lon_var,
  clinical_id_var,
  exposures = NULL,
  exposures_id_var = NULL
)
}
\arguments{
\item{nc}{}

\item{varid}{}
}
\description{
Title
}
