% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputationDS.R
\name{imputationDS}
\alias{imputationDS}
\title{Perform a imputation of an Exposome Set}
\usage{
imputationDS(object, select = NULL)
}
\arguments{
\item{object}{\code{ExposomeSet} Exposome Set object}

\item{select}{\code{character} \code{character} (default \code{NULL}) Name to be assigned to the exposed set on the server side. If missing,
the input Exposome Set will be over written.}
}
\value{
\code{ExposomeSet} Imputed Exposome Set
}
\description{
Perform a imputation of an Exposome Set using rexposome base function
}
