% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invExWASDS.R
\name{invExWASDS}
\alias{invExWASDS}
\title{Title}
\usage{
invExWASDS(object, phenotype, tef, ...)
}
\arguments{
\item{object}{\code{ExposomeSet} Exposome object}

\item{phenotype}{\code{character} Phenotype objective}

\item{tef}{\code{bool} If \code{TRUE} computes the threshold for effective tests.}

\item{...}{\code{character vector} Adjusting phenotype covariables}
}
\value{
\code{ExWAS} object. With tef, coefficients and pvalues of associations
}
\description{
Title
}
