% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadExposomeDS.R
\name{loadExposomeDS}
\alias{loadExposomeDS}
\title{Create an ExposomeSet from \code{data.frames}}
\usage{
loadExposomeDS(
  exposures,
  description,
  phenotype,
  description.famCol = "family",
  exposures.asFactor = 5,
  warnings = TRUE
)
}
\arguments{
\item{exposures}{\code{data.frame} With the exposures of the exposome}

\item{description}{\code{data.frame} With the description of the exposome}

\item{phenotype}{\code{data.frame} With the phenotype of the exposome}

\item{description.famCol}{(default \code{"family"}) Index where the family's
name (per exposures) if found in file "description". It can be both numeric
or character.}

\item{exposures.asFactor}{\code{numeric} (default \code{5}) The exposures with more
than this number of unique items will be considered as "continuous" while
the exposures with less or equal number of items will be considered as
"factor".}

\item{warnings}{\code{bool} (default \code{FALSE}) If \code{TRUE} shows useful}
}
\value{
An \code{ExposomeSet} object
}
\description{
Takes the three tables of an exposome dataset and coerces them into a Exposome Set object
on the study server
}
