% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturityoffsetDS.R
\name{maturityoffsetDS}
\alias{maturityoffsetDS}
\title{Maturity offset estimation (pubertal development)}
\usage{
maturityoffsetDS(data, sex, age, height)
}
\arguments{
\item{data}{Dataset name}

\item{sex}{Name of the sex variable in the input dataset (0 for males and 1 for females)}

\item{age}{Name of the Age variable in the input dataset (in days)}

\item{height}{Name of the Height variable in the input dataset (in centimeters)}
}
\value{
Estimated maturity offset according to Moore's equations.
}
\description{
Calculate maturity offset according to the equations of Moore et al. 2015 (10.1249/MSS.0000000000000588)
}
\details{
More details can be found at Moore et al. 2015 (10.1249/MSS.0000000000000588)
}
\author{
Augusto Anguita-Ruiz & Xavier Escribá
}
