% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetCDF_fillvalue_matrixDS.R
\name{NetCDF_fillvalue_matrixDS}
\alias{NetCDF_fillvalue_matrixDS}
\title{Set NAs to matrix fillvalues}
\usage{
NetCDF_fillvalue_matrixDS(mat, fillvalue)
}
\arguments{
\item{mat}{`matrix` Extracted from a `NetCDF` resource using `nc_dataDS`}

\item{fillvalue}{`list` Extracted from the `_FillValue` slot of a `NetCDF`
resource using `ncatt_getDS`}
}
\value{
`matrix`
}
\description{
Use a fillvalue extracted using `ncatt_get` and place NAs
}
