% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcdf_varsDS.R
\name{netcdf_varsDS}
\alias{netcdf_varsDS}
\title{Get variable names of NetCDF object}
\usage{
netcdf_varsDS(nc)
}
\arguments{
\item{nc}{An object of class ncdf4 (loaded from a NetCDF resource), 
indicating what file to get the variables from.}
}
\value{
`character vector` With the variable names
}
\description{
Get variable names of NetCDF object
}
