% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetCDFFileResourceClient.R
\docType{class}
\name{NetCDFFileResourceClient}
\alias{NetCDFFileResourceClient}
\title{NetCDF file resource client}
\format{
A R6 object of class NetCDFFileResourceClient
}
\description{
Build a NetCDF handler from a resource object describing access to a 
NetCDF file to be studied.
}
\section{Super classes}{
\code{\link[resourcer:ResourceClient]{resourcer::ResourceClient}} -> \code{\link[resourcer:FileResourceClient]{resourcer::FileResourceClient}} -> \code{NetCDFFileResourceClient}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NetCDFFileResourceClient-new}{\code{NetCDFFileResourceClient$new()}}
\item \href{#method-NetCDFFileResourceClient-getConnection}{\code{NetCDFFileResourceClient$getConnection()}}
\item \href{#method-NetCDFFileResourceClient-getValue}{\code{NetCDFFileResourceClient$getValue()}}
\item \href{#method-NetCDFFileResourceClient-close}{\code{NetCDFFileResourceClient$close()}}
\item \href{#method-NetCDFFileResourceClient-clone}{\code{NetCDFFileResourceClient$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="asDataFrame"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-asDataFrame'><code>resourcer::ResourceClient$asDataFrame()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="asTbl"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-asTbl'><code>resourcer::ResourceClient$asTbl()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="exec"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-exec'><code>resourcer::ResourceClient$exec()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="getResource"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-getResource'><code>resourcer::ResourceClient$getResource()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="FileResourceClient" data-id="downloadFile"><a href='../../resourcer/html/FileResourceClient.html#method-FileResourceClient-downloadFile'><code>resourcer::FileResourceClient$downloadFile()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetCDFFileResourceClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-NetCDFFileResourceClient-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetCDFFileResourceClient$new(resource)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetCDFFileResourceClient-getConnection"></a>}}
\if{latex}{\out{\hypertarget{method-NetCDFFileResourceClient-getConnection}{}}}
\subsection{Method \code{getConnection()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetCDFFileResourceClient$getConnection(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetCDFFileResourceClient-getValue"></a>}}
\if{latex}{\out{\hypertarget{method-NetCDFFileResourceClient-getValue}{}}}
\subsection{Method \code{getValue()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetCDFFileResourceClient$getValue(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetCDFFileResourceClient-close"></a>}}
\if{latex}{\out{\hypertarget{method-NetCDFFileResourceClient-close}{}}}
\subsection{Method \code{close()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetCDFFileResourceClient$close()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetCDFFileResourceClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NetCDFFileResourceClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetCDFFileResourceClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
