% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetCDFFileResourceResolver.R
\docType{class}
\name{NetCDFFileResourceResolver}
\alias{NetCDFFileResourceResolver}
\title{NetCDF file resource resolver}
\format{
A R6 object of class NetCDFFileResourceResolver
}
\description{
Build a NetCDF resource client from a resource object describing access to a 
NetCDF file to be studied.
}
\section{Super class}{
\code{\link[resourcer:ResourceResolver]{resourcer::ResourceResolver}} -> \code{NetCDFFileResourceResolver}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NetCDFFileResourceResolver-isFor}{\code{NetCDFFileResourceResolver$isFor()}}
\item \href{#method-NetCDFFileResourceResolver-newClient}{\code{NetCDFFileResourceResolver$newClient()}}
\item \href{#method-NetCDFFileResourceResolver-clone}{\code{NetCDFFileResourceResolver$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceResolver" data-id="initialize"><a href='../../resourcer/html/ResourceResolver.html#method-ResourceResolver-initialize'><code>resourcer::ResourceResolver$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetCDFFileResourceResolver-isFor"></a>}}
\if{latex}{\out{\hypertarget{method-NetCDFFileResourceResolver-isFor}{}}}
\subsection{Method \code{isFor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetCDFFileResourceResolver$isFor(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetCDFFileResourceResolver-newClient"></a>}}
\if{latex}{\out{\hypertarget{method-NetCDFFileResourceResolver-newClient}{}}}
\subsection{Method \code{newClient()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetCDFFileResourceResolver$newClient(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NetCDFFileResourceResolver-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NetCDFFileResourceResolver-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetCDFFileResourceResolver$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
