% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalityTestDS.R
\name{normalityTestDS}
\alias{normalityTestDS}
\title{Normality test of exposures of ExposomeSet}
\usage{
normalityTestDS(object, th = 0.05, min.val = 5, na.rm = TRUE, warnings = TRUE)
}
\arguments{
\item{object}{\code{ExposomeSet} Exposome Set object}

\item{th}{\code{numeric} (default \code{0.05}) Threshold to considere an exposure to follow a normal distribution.}

\item{min.val}{\code{numeric} (default \code{5}) Minimum number of values not missings to test the exposures.}

\item{na.rm}{\code{bool} (default \code{TRUE}) Removes the NA values to test the normality on the exposure.}

\item{warnings}{\code{bool} (default \code{TRUE}) Show warnings if required.}
}
\value{
list of \code{data.frame} (one for each study server) with \cr
- exposure: \code{character} Name of the exposures \cr
- normality: \code{bool} If exposure is normal (\code{TRUE}) or not (\code{FALSE}) \cr
- p.value: \code{numeric} P-value of the Shapiro-Wilk Normality Test
}
\description{
Perform a normality test on all the exposures of an ExposomeSet object
}
