% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypeNamesDS.R
\name{phenotypeNamesDS}
\alias{phenotypeNamesDS}
\title{Exposure Phenotype names}
\usage{
phenotypeNamesDS(exposomeSet)
}
\arguments{
\item{exposomeSet}{\code{ExposomeSet} ExposomeSet object}
}
\value{
\code{character vector} of the phenotypes names
}
\description{
Extract the phenotypes names from a Expression Set object
}
