% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFamilyDS.R
\name{plotFamilyDS}
\alias{plotFamilyDS}
\title{Draw boxplot of a family of an Exposome Set}
\usage{
plotFamilyDS(
  x,
  family,
  group = NA,
  group2 = NA,
  scatter = TRUE,
  na.omit = TRUE
)
}
\arguments{
\item{x}{\code{ExposomeSet} Exposome Set object}

\item{family}{\code{character} Name of the familty that will be drawn}

\item{group}{\code{character} If set it displays the family grouped by the given phenotype}

\item{group2}{\code{character} If set it displays the family grouped by the given phenotype}

\item{scatter}{\code{bool} (default \code{TRUE}) If the family to be plotted is continuous, the samples will be shown}

\item{na.omit}{\code{bool} (default \code{TRUE}) Do not show NA values}
}
\value{
Returns a ggplot object
}
\description{
Draw boxplot of a family of an Exposome Set
}
