% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMissingsDS.R
\name{plotMissingsDS}
\alias{plotMissingsDS}
\title{Draw plot with percentage of missings}
\usage{
plotMissingsDS(exp, set = "exposures")
}
\arguments{
\item{exp}{\code{ExposomeSet} Exposome Set object}

\item{set}{\code{character} (default \code{"exposures"}) Set to get the missings plot: \code{"exposures"} or \code{"phenotypes"}}
}
\value{

}
\description{
Create a plot with the percentage of missings for the exposures and phenotypes of an 
ExposomeSet object
}
