% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleDS.R
\name{scaleDS}
\alias{scaleDS}
\title{Scale a data frame}
\usage{
scaleDS(x, center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{\code{data frame} Table to be scaled}

\item{center}{Corresponds to the \code{center} parameter of the base function \code{scale}}

\item{scale}{Corresponds to the \code{scale} parameter of the base function \code{scale}}
}
\value{
\code{data frame} Scaled table
}
\description{
Scale a data frame
}
