% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapiro.testDS.R
\name{shapiro.testDS}
\alias{shapiro.testDS}
\title{Shapiro-Wilk Normality Test}
\usage{
shapiro.testDS(x)
}
\arguments{
\item{x}{\code{numeric} Name of the numeric vector}
}
\value{
Results of the Shapiro-Wilk Normality Test
}
\description{
Shapiro-Wilk Normality Test
}
