% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposome_pca_pooledDS.R
\name{svdPartial}
\alias{svdPartial}
\title{Calculate partial SVD}
\usage{
svdPartial(x)
}
\arguments{
\item{x}{\code{matrix}}
}
\value{
\code{matrix} with results
}
\description{
Calculate partial SVD
}
