% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMissingsDS.R
\name{tableMissingsDS}
\alias{tableMissingsDS}
\title{Get table of missings per variable}
\usage{
tableMissingsDS(exp, set = "exposures", output = "n")
}
\arguments{
\item{exp}{\code{ExposomeSet} Exposome Set object}

\item{set}{\code{character} (default \code{"exposures"}) Set to get the missings: \code{"exposures"} or \code{"phenotypes"}}

\item{output}{\code{character} (default \code{"n"}) Get missing number (\code{"n"}) or percentage (\code{"p"})}
}
\value{
Vector of named numerics, where the name corresponds to the variable and the associated numeric corresponds
to the missing (number or percentage, dependeing on the \code{output} parameter)
}
\description{
Obtain the number (or percentage) of missings of for the exposures or phenotypes of an ExposomeSet object
}
