% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformDS.R
\name{transformDS}
\alias{transformDS}
\title{Apply a transformation to the exposure of an ExposomeSet}
\usage{
transformDS(Set, exposure, method)
}
\arguments{
\item{Set}{`ExposomeSet` which exposures will be transformed}

\item{exposure}{`character` Name of the exposure to be transformed}

\item{method}{`fun` Function to be applied to the exposure}
}
\value{
`ExposomeSet` with a transformed exposure
}
\description{
Apply a transformation to the exposure of an ExposomeSet
}
