% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposome_pca_plotDS.R
\name{exposome_pca_plotDS}
\alias{exposome_pca_plotDS}
\title{Title}
\usage{
exposome_pca_plotDS(object, set, phenotype, method, k, noise)
}
\arguments{
\item{object}{}

\item{set}{}

\item{phenotype}{}
}
\value{

}
\description{
Title
}
