% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposome_pca_pooledDS.R
\name{exposome_pca_pooled_addPCDS}
\alias{exposome_pca_pooled_addPCDS}
\title{Add PCA results to ExposomeSet}
\usage{
exposome_pca_pooled_addPCDS(object, pca)
}
\arguments{
\item{object}{\code{ExposomeSet} to add the PCA results}

\item{pca}{\code{raw} Serialized PCA object}
}
\value{
\code{ExposomePCA}
}
\description{
Adds the results of the block method to the ExposomeSet, creating a new ExposomePCA.
}
