% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposome_pcaDS.R
\name{exposome_pcaDS}
\alias{exposome_pcaDS}
\title{Principal components analysis of an Exposome Set}
\usage{
exposome_pcaDS(object, npc, pca)
}
\arguments{
\item{object}{\code{ExposomeSet} Exposome Set object}

\item{npc}{\code{numeric} Number of PC to be kept}

\item{pca}{\code{bool} If TRUE perform PCA (only numerical variables), if FALSE FAMD (numerical and categorical)}
}
\value{
Returns a \code{ExposomeSet PCA} object
}
\description{
Performs a non-disclosive PCA given an Exposome Set, the Exposome Set can be subsetted by families to 
perform the PCA
}
