% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposures_pData.R
\name{exposures_pData}
\alias{exposures_pData}
\title{Exposures and phenotyphes data}
\usage{
exposures_pData(exposomeSet, target = "all")
}
\arguments{
\item{exposomeSet}{\code{ExposomeSet} Exposome Set object}

\item{target}{\code{character} To specify target of output table, \code{all} to include exposures and phenotypes,
\code{exposures} to include only the exposures and \code{phenotypes} to include only the phenotypes. Default \code{all}}
}
\value{
\code{data.frame} With exposures and phenotypes
}
\description{
Extract a table with the exposures and phenotypes data for all the individuals of an
Exposome Set
}
