% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractExposomeSetFromImputedSetDS.R
\name{extractExposomeSetFromImputedSetDS}
\alias{extractExposomeSetFromImputedSetDS}
\title{Title}
\usage{
extractExposomeSetFromImputedSetDS(x, imputation)
}
\arguments{
\item{x}{}

\item{imputation}{}
}
\value{

}
\description{
Title
}
