% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetCDF_fillvalue_matrixDS.R
\name{NetCDF_fillvalue_matrixDS}
\alias{NetCDF_fillvalue_matrixDS}
\title{Title}
\usage{
NetCDF_fillvalue_matrixDS(mat, fillvalue)
}
\arguments{
\item{nc}{}

\item{varid}{}
}
\description{
Title
}
