% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFamilyDS.R
\name{plotFamilyDS}
\alias{plotFamilyDS}
\title{Draw boxplot of a family of an Exposome Set}
\usage{
plotFamilyDS(
  x,
  family,
  group = NA,
  group2 = NA,
  scatter = TRUE,
  na.omit = TRUE,
  method,
  k,
  noise
)
}
\arguments{
\item{x}{\code{ExposomeSet} Exposome Set object}

\item{family}{\code{character} Name of the familty that will be drawn}

\item{group}{\code{character} If set it displays the family grouped by the given phenotype}

\item{group2}{\code{character} If set it displays the family grouped by the given phenotype}

\item{scatter}{\code{bool} (default \code{TRUE}) If the family to be plotted is continuous, the samples will be shown}

\item{na.omit}{\code{bool} (default \code{TRUE}) Do not show NA values}

\item{method}{\code{numeric} either 1 or 2. If the user selects the deterministic method in the 
client side function the method.inticator is set to 1 while if the user selects the probabilistic method this argument is set to 2.}

\item{k}{\code{numeric} the number of the nearest neghbours for which their centroid is calculated 
if the deterministic method is selected.}

\item{noise}{\code{numeric} the percentage of the initial variance that is used as the variance of 
the embedded noise if the probabilistic method is selected.}
}
\value{
Returns a ggplot object
}
\description{
Draw boxplot of a family of an Exposome Set
}
